requirements.txt
python_deriv_api
tensorflow
colorama
pandas
numpy
python-dotenv
InquirerPy